Overhauled the LogitechMic class to match PCSX2 (and by extension QEMU) a bit closer on its implementation.
So far, all the titles I've tested seem to have all of its CtrlMessage transfers handled fine; it even seems to properly set the interface and sample rate (usually to 48kHz) when the game requests it. What's yet to be done is getting a signal to transmit through; titles will acknowledge the microphone but it doesn't seem like there's any response just yet.